Process videos in series
File system has a limit on the number of concurrent accesses it can have an this causes failures

Linear processing means we hit the 5 minute execution limit. Since this job will run as a batch, this means the next run will continue with missing transcripts. No data is lost.

Parallel processing with concurrency would have been a good solution but it does not work.