Bot should reply to thread
Bring back keep_warm=1 because slack sends the message repeatedly if we don't respond in 3 seconds. This means we get multiple responses (up to 3 if the timeout is not hit)

Start caching embeddings to improve performance

thread_ts is an optional field (not available when the message isn't in a thread.